+++ routes/reviewRoutes.js
@@ -2,6 +2,7 @@ import express from "express";
 import listingReview from "../config_DB/models/listingReviewSchema.js";
 import { authorizationCheck } from "../customMiddlewares.js";
 import { reviewValidation } from "../customMiddlewares.js";
+import ExpressError from "../error.js";
 
 const router = express.Router();
 
@@ -10,24 +11,19 @@ router.post(
   "/:id/listing",
   reviewValidation,
   authorizationCheck,
-  async (req, res) => {
+  async (req, res, next) => {
     const reviewData = req.body.listingReview;
     try {
       await listingReview.create({
         ...reviewData,
         author: req.session.userID,
       });
+      req.flash("success", "Review posted");
       return res.redirect(`/listing/${req.params.id}`);
     } catch (reviewError) {
       console.log(reviewError);
-      res.render("Error", {
-        reviewError,
-        error: null,
-        error1: null,
-        invalidID: null,
-        idNotFound: null,
-        userError: null,
-      });
+      const newErr = new ExpressError("Review couldnt be posted", 500);
+      next(newErr);
     }
   },
 );
@@ -38,22 +34,29 @@ router.get("/new/:id", authorizationCheck, (req, res) => {
 });
 
 // delete review
-router.post(
+router.delete(
   "/delete/:reviewId/:listingID",
   authorizationCheck,
-  async (req, res) => {
+  async (req, res, next) => {
     const reviewID = req.params.reviewId;
     const listingID = req.params.listingID;
     try {
-      await listingReview.findByIdAndDelete(reviewID);
-      res.redirect(`/listing/${listingID}`);
+      const result = await listingReview.findById({ _id: reviewID });
+      if (!result) {
+        res.send("Review not found unable to delete").status(400);
+        return;
+      }
+      if (result.author.equals(req.session.userID)) {
+        await listingReview.findByIdAndDelete(reviewID);
+        req.flash("success", "Review deleted successfully");
+        res.redirect(`/listing/${listingID}`);
+      } else {
+        res.send("Cannot delete review you are not the author");
+      }
     } catch (deleteReviewError) {
       console.log(deleteReviewError);
-      res.render("Error", {
-        deleteReviewError,
-        error: null,
-        reviewError: null,
-      });
+      const newErr = new ExpressError("Couldnt delete the review", 500);
+      next(newErr);
     }
   },
 );
